move update_activity_with_track to verve_backend.api.common.utils
